requirements.txt
-e .
aws-cdk.aws-lambda==1.129.0